PistonCrafting with Tags
moved PistonCrafting to own world/Pistoncrafting class
Added Tags to SushiContent
Created pressing_base tag
modified Pistoncrafting to check for Block Tag
