+++ src/main/resources/data/sushigocrafting/tags/block/pressing_base.json
@@ -1,0 +1,5 @@
+{
+  "values": [
+    "minecraft:cobblestone"
+  ]
+}
